+++ WebSocket4Net/Protocol/ICloseStatusCode.cs
@@ -4,16 +4,16 @@ namespace WebSocket4Net.Protocol
 {
     public interface ICloseStatusCode
     {
-        int ExtensionNotMatch { get; }
-        int GoingAway { get; }
-        int InvalidUTF8 { get; }
-        int NormalClosure { get; }
-        int NotAcceptableData { get; }
-        int ProtocolError { get; }
-        int TLSHandshakeFailure { get; }
-        int TooLargeFrame { get; }
-        int UnexpectedCondition { get; }
-        int ViolatePolicy { get; }
-        int NoStatusCode { get; }
+        short ExtensionNotMatch { get; }
+        short GoingAway { get; }
+        short InvalidUTF8 { get; }
+        short NormalClosure { get; }
+        short NotAcceptableData { get; }
+        short ProtocolError { get; }
+        short TLSHandshakeFailure { get; }
+        short TooLargeFrame { get; }
+        short UnexpectedCondition { get; }
+        short ViolatePolicy { get; }
+        short NoStatusCode { get; }
     }
 }
